Move flux points and estimator tests
