+++ gammapy/estimators/points/__init__.py
@@ -2,3 +2,4 @@ from .lightcurve import *
 from .sensitivity import *
 from .core import *
 from .profile import *
+from .sed import *
