+++ gammapy/estimators/points/core.py
@@ -2,27 +2,22 @@
 import logging
 import numpy as np
 from scipy import stats
-from astropy import units as u
 from astropy.io.registry import IORegistryError
 from astropy.table import Table, vstack
 from astropy.visualization import quantity_support
-from gammapy.datasets import Datasets
 from gammapy.modeling.models import TemplateSpectralModel
 from gammapy.modeling.models.spectral import scale_plot_flux
-from gammapy.modeling import Fit
 from gammapy.maps import RegionNDMap, Maps, TimeMapAxis, MapAxis
 from gammapy.maps.axes import flat_if_equal
 from gammapy.utils.scripts import make_path
-from gammapy.utils.pbar import progress_bar
-from gammapy.utils.table import table_from_row_data, table_standardise_units_copy
+from gammapy.utils.table import table_standardise_units_copy
 from ..map.core import (
     FluxMaps,
     DEFAULT_UNIT,
 )
-from ..flux import FluxEstimator
 
 
-__all__ = ["FluxPoints", "FluxPointsEstimator"]
+__all__ = ["FluxPoints"]
 
 log = logging.getLogger(__name__)
 
@@ -590,169 +585,3 @@ class FluxPoints(FluxMaps):
             ax.figure.colorbar(caxes, ax=ax, label=label)
 
         return ax
-
-
-class FluxPointsEstimator(FluxEstimator):
-    """Flux points estimator.
-
-    Estimates flux points for a given list of datasets, energies and spectral model.
-
-    To estimate the flux point the amplitude of the reference spectral model is
-    fitted within the energy range defined by the energy group. This is done for
-    each group independently. The amplitude is re-normalized using the "norm" parameter,
-    which specifies the deviation of the flux from the reference model in this
-    energy group. See https://gamma-astro-data-formats.readthedocs.io/en/latest/spectra/binned_likelihoods/index.html
-    for details.
-
-    The method is also described in the Fermi-LAT catalog paper
-    https://ui.adsabs.harvard.edu/#abs/2015ApJS..218...23A
-    or the HESS Galactic Plane Survey paper
-    https://ui.adsabs.harvard.edu/#abs/2018A%26A...612A...1H
-
-    Parameters
-    ----------
-    energy_edges : `~astropy.units.Quantity`
-        Energy edges of the flux point bins.
-    source : str or int
-        For which source in the model to compute the flux points.
-    norm_min : float
-        Minimum value for the norm used for the fit statistic profile evaluation.
-    norm_max : float
-        Maximum value for the norm used for the fit statistic profile evaluation.
-    norm_n_values : int
-        Number of norm values used for the fit statistic profile.
-    norm_values : `numpy.ndarray`
-        Array of norm values to be used for the fit statistic profile.
-    n_sigma : int
-        Number of sigma to use for asymmetric error computation. Default is 1.
-    n_sigma_ul : int
-        Number of sigma to use for upper limit computation. Default is 2.
-    selection_optional : list of str
-        Which additional quantities to estimate. Available options are:
-
-            * "all": all the optional steps are executed
-            * "errn-errp": estimate asymmetric errors on flux.
-            * "ul": estimate upper limits.
-            * "scan": estimate fit statistic profiles.
-
-        Default is None so the optionnal steps are not executed.
-    fit : `Fit`
-        Fit instance specifying the backend and fit options.
-    reoptimize : bool
-        Re-optimize other free model parameters. Default is True.
-    """
-
-    tag = "FluxPointsEstimator"
-
-    def __init__(self, energy_edges=[1, 10] * u.TeV, **kwargs):
-        self.energy_edges = energy_edges
-
-        fit = Fit(confidence_opts={"backend": "scipy"})
-        kwargs.setdefault("fit", fit)
-        super().__init__(**kwargs)
-
-    def run(self, datasets):
-        """Run the flux point estimator for all energy groups.
-
-        Parameters
-        ----------
-        datasets : `~gammapy.datasets.Datasets`
-            Datasets
-
-        Returns
-        -------
-        flux_points : `FluxPoints`
-            Estimated flux points.
-        """
-        datasets = Datasets(datasets=datasets)
-        rows = []
-
-        for energy_min, energy_max in progress_bar(
-            zip(self.energy_edges[:-1], self.energy_edges[1:]), desc="Energy bins"
-        ):
-            row = self.estimate_flux_point(
-                datasets,
-                energy_min=energy_min,
-                energy_max=energy_max,
-            )
-            print(row)
-            rows.append(row)
-
-        meta = {
-            "n_sigma": self.n_sigma,
-            "n_sigma_ul": self.n_sigma_ul,
-            "sed_type_init": "likelihood",
-        }
-
-        table = table_from_row_data(rows=rows, meta=meta)
-        model = datasets.models[self.source]
-        return FluxPoints.from_table(
-            table=table,
-            reference_model=model.copy(),
-            gti=datasets.gti,
-            format="gadf-sed",
-        )
-
-    def estimate_flux_point(self, datasets, energy_min, energy_max):
-        """Estimate flux point for a single energy group.
-
-        Parameters
-        ----------
-        datasets : `Datasets`
-            Datasets
-        energy_min, energy_max : `~astropy.units.Quantity`
-            Energy bounds to compute the flux point for.
-
-        Returns
-        -------
-        result : dict
-            Dict with results for the flux point.
-        """
-        datasets_sliced = datasets.slice_by_energy(
-            energy_min=energy_min, energy_max=energy_max
-        )
-
-        if len(datasets_sliced) > 0:
-            datasets_sliced.models = datasets.models.copy()
-            return super().run(datasets=datasets_sliced)
-        else:
-            log.warning(f"No dataset contribute in range {energy_min}-{energy_max}")
-            model = datasets.models[self.source].spectral_model
-            return self._nan_result(datasets, model, energy_min, energy_max)
-
-    def _nan_result(self, datasets, model, energy_min, energy_max):
-        """NaN result"""
-        energy_axis = MapAxis.from_energy_edges([energy_min, energy_max])
-
-        with np.errstate(invalid="ignore", divide="ignore"):
-            result = model.reference_fluxes(energy_axis=energy_axis)
-            # convert to scalar values
-            result = {key: value.item() for key, value in result.items()}
-
-        result.update(
-            {
-                "norm": np.nan,
-                "stat": np.nan,
-                "success": False,
-                "norm_err": np.nan,
-                "ts": np.nan,
-                "counts": np.zeros(len(datasets)),
-                "npred": np.nan * np.zeros(len(datasets)),
-                "npred_excess": np.nan * np.zeros(len(datasets)),
-                "datasets": datasets.names,
-            }
-        )
-
-        if "errn-errp" in self.selection_optional:
-            result.update({"norm_errp": np.nan, "norm_errn": np.nan})
-
-        if "ul" in self.selection_optional:
-            result.update({"norm_ul": np.nan})
-
-        if "scan" in self.selection_optional:
-            norm = super()._set_norm_parameter()
-            norm_scan = norm.scan_values
-            result.update({"norm_scan": norm_scan, "stat_scan": np.nan * norm_scan})
-
-        return result
-
