+++ gammapy/estimators/points/tests/test_core.py
